Revert unneeded SecretsRetriever changes
