--- Lambdas/LambdasShared/GenericSecretsRetriever.swift
@@ -1,3 +1,0 @@
-package protocol GenericSecretsRetriever: Sendable {
-    func getSecret(arnEnvVarKey: String) async throws -> String
-}
