--- Tests/PennyTests/Fake/FakeSecretsRetriever.swift
@@ -1,7 +1,0 @@
-@testable import LambdasShared
-
-struct FakeSecretsRetriever: GenericSecretsRetriever {
-    func getSecret(arnEnvVarKey: String) async throws -> String {
-        "test-secret"
-    }
-}
